move the stuff back into index route
