+++ app/root.tsx
@@ -1,35 +1,22 @@
 import {
   Deferrable,
   deferred,
+  ErrorBoundaryComponent,
   LinksFunction,
   LoaderFunction,
   MetaFunction,
 } from "@remix-run/node"
 import {
-  Deferred,
   Links,
   LiveReload,
   Meta,
+  Outlet,
   Scripts,
   useCatch,
-  useLoaderData,
-  useNavigate,
-  useSearchParams,
 } from "@remix-run/react"
 import type { CatchBoundaryComponent } from "@remix-run/react/routeModules"
-import { Virtuoso } from "react-virtuoso"
 import { FontDict, loadFonts } from "~/modules/fonts/api.server"
-import { FontCard } from "~/modules/fonts/font-list"
-import { SearchForm } from "~/modules/ui/search-form"
-import { useWindowSize } from "./modules/dom/use-window-size"
-import { ClearSelectedFontsButton } from "./modules/fonts/clear-selected-fonts-button"
-import { makeFontContext } from "./modules/fonts/font-context"
 import { pangrams } from "./modules/fonts/pangrams"
-import { SaveFontsButton } from "./modules/fonts/save-fonts-button"
-import { SelectedFont } from "./modules/fonts/selected-font"
-import { makeSearchContext } from "./modules/search/search-context"
-import { MaxWidthContainer } from "./modules/ui/max-width-container"
-import { RaisedPanel } from "./modules/ui/raised-panel"
 import tailwind from "./tailwind.css"
 
 type LoaderData = {
@@ -75,106 +62,22 @@ function Document({ children }: { children: React.ReactNode }) {
 }
 
 export default function App() {
-  const { fonts, pangrams } = useLoaderData<LoaderData>()
-  const [params] = useSearchParams()
-  const searchContext = makeSearchContext(params)
-  const navigate = useNavigate()
-  const { height } = useWindowSize()
-
   return (
     <Document>
-      <LeftSidebar>
-        <RaisedPanel rounded={false} fullHeight>
-          <Deferred value={fonts} fallback={<p>Loading...</p>}>
-            {(fonts) => {
-              const context = makeFontContext(fonts, params)
-              return (
-                <section
-                  aria-label="Font selections"
-                  className="flex flex-col h-full gap-4 p-4 w-80"
-                >
-                  {context.selectedFontList.map((font) => (
-                    <SelectedFont
-                      key={font.family}
-                      font={font}
-                      context={context}
-                    />
-                  ))}
-                  <div className="flex-1" />
-                  <ClearSelectedFontsButton context={context} />
-                  <SaveFontsButton />
-                </section>
-              )
-            }}
-          </Deferred>
-        </RaisedPanel>
-      </LeftSidebar>
-
-      <div className="pl-80 -z-10">
-        <header className="sticky top-0 z-10 py-4 shadow-md bg-base-200">
-          <MaxWidthContainer>
-            <Header>
-              <SearchForm
-                name={searchContext.paramName}
-                defaultValue={searchContext.searchQuery}
-                onSubmit={(value) => {
-                  navigate(searchContext.getSearchLink(value), {
-                    replace: true,
-                  })
-                }}
-              />
-            </Header>
-          </MaxWidthContainer>
-        </header>
-
-        <main className="py-4">
-          <MaxWidthContainer>
-            <Deferred
-              value={fonts}
-              fallback={
-                <p className="p-3 text-center opacity-50">Loading...</p>
-              }
-            >
-              {(fonts) => {
-                const context = makeFontContext(fonts, params)
-                return (
-                  <div className="-my-2">
-                    <Virtuoso
-                      useWindowScroll
-                      data={context.fontList}
-                      overscan={height}
-                      itemContent={(index, font) => (
-                        <div className="py-2">
-                          <FontCard
-                            font={font}
-                            context={context}
-                            previewText={pangrams[index % pangrams.length]!}
-                          />
-                        </div>
-                      )}
-                    />
-                  </div>
-                )
-              }}
-            </Deferred>
-          </MaxWidthContainer>
-        </main>
-      </div>
+      <Outlet />
     </Document>
   )
 }
 
-function Header({ children }: { children: React.ReactNode }) {
+export const ErrorBoundary: ErrorBoundaryComponent = ({ error }) => {
   return (
-    <nav className="flex items-center gap-6">
-      <h1 className="text-3xl">fontme</h1>
-      <div className="flex-1">{children}</div>
-    </nav>
+    <Document>
+      <main>
+        <h1>oops, something went wrong</h1>
+        <pre className="overflow-x-auto">{error.stack}</pre>
+      </main>
+    </Document>
   )
-}
-
-function LeftSidebar({ children }: { children: React.ReactNode }) {
-  return <div className="fixed inset-y-0 left-0">{children}</div>
 }
 
 export const CatchBoundary: CatchBoundaryComponent = (props) => {
